# Conflicts:
#	activitysim/core/pipeline.py

+++ activitysim/core/pipeline.py
@@ -4,6 +4,7 @@ import datetime as dt
 import logging
 import os
 from builtins import map, next
+from pathlib import Path
 
 import pandas as pd
 from pypyr.context import Context
@@ -57,7 +58,7 @@ class Pipeline:
         self.context = Context()
         self.init_state()
 
-    def init_state(self):
+    def init_state(self, pipeline_file_format="parquet"):
 
         # most recent checkpoint
         self.last_checkpoint = {}
@@ -93,7 +94,7 @@ class Pipeline:
     def is_readonly(self):
         if self.is_open:
             store = self.get_pipeline_store()
-            if store and store._mode == "r":
+            if store and not isinstance(store, Path) and store._mode == "r":
                 return True
         return False
 
@@ -116,7 +117,11 @@ class Pipeline:
 
     def open_pipeline_store(self, overwrite=False, mode="a"):
         """
-        Open the pipeline checkpoint store
+        Open the pipeline checkpoint store.
+
+        If the pipeline_file_name setting ends in ".h5", then the pandas
+        HDFStore file format is used, otherwise pipeline files are stored
+        as parquet files organized in regular file system directories.
 
         Parameters
         ----------
@@ -142,22 +147,30 @@ class Pipeline:
             inject.get_injectable("pipeline_file_name")
         )
 
-        if overwrite:
-            try:
-                if os.path.isfile(pipeline_file_path):
-                    logger.debug("removing pipeline store: %s" % pipeline_file_path)
-                    os.unlink(pipeline_file_path)
-            except Exception as e:
-                print(e)
-                logger.warning("Error removing %s: %s" % (pipeline_file_path, e))
+        if pipeline_file_path.endswith(".h5"):
+            if overwrite:
+                try:
+                    if os.path.isfile(pipeline_file_path):
+                        logger.debug("removing pipeline store: %s" % pipeline_file_path)
+                        os.unlink(pipeline_file_path)
+                except Exception as e:
+                    print(e)
+                    logger.warning("Error removing %s: %s" % (pipeline_file_path, e))
 
-        self.pipeline_store = pd.HDFStore(pipeline_file_path, mode=mode)
+            self.pipeline_store = pd.HDFStore(pipeline_file_path, mode=mode)
+        else:
+            self.pipeline_store = Path(pipeline_file_path)
 
         logger.debug(f"opened pipeline_store {pipeline_file_path}")
 
     def get_pipeline_store(self):
         """
         Return the open pipeline hdf5 checkpoint store or return None if it not been opened
+
+        If the pipeline filename ends in ".h5" then the legacy HDF5 pipeline
+        is used, otherwise the faster parquet format is used, and the value
+        returned here is just the path to the pipeline directory.
+
         """
         return self.pipeline_store
 
@@ -195,7 +208,12 @@ class Pipeline:
         """
 
         store = self.get_pipeline_store()
-        df = store[self.pipeline_table_key(table_name, checkpoint_name)]
+        if isinstance(store, Path):
+            df = pd.read_parquet(
+                store.joinpath(table_name, f"{checkpoint_name}.parquet"),
+            )
+        else:
+            df = store[self.pipeline_table_key(table_name, checkpoint_name)]
 
         return df
 
@@ -206,7 +224,10 @@ class Pipeline:
         We store multiple versions of all simulation tables, for every checkpoint in which they change,
         so we need to know both the table_name and the checkpoint_name to label the saved table
 
-        The only exception is the checkpoints dataframe, which just has a table_name
+        The only exception is the checkpoints dataframe, which just has a table_name,
+        although when using the parquet storage format this file is stored as "None.parquet"
+        to maintain a simple consistent file directory structure.
+
 
         Parameters
         ----------
@@ -222,10 +243,28 @@ class Pipeline:
         df.columns = df.columns.astype(str)
 
         store = self.get_pipeline_store()
-
-        store[self.pipeline_table_key(table_name, checkpoint_name)] = df
-
-        store.flush()
+        if isinstance(store, Path):
+            store.joinpath(table_name).mkdir(parents=True, exist_ok=True)
+            df.to_parquet(store.joinpath(table_name, f"{checkpoint_name}.parquet"))
+        else:
+            complib = config.setting("pipeline_complib", None)
+            if complib is None or len(df.columns) == 0:
+                # tables with no columns can't be compressed successfully, so to
+                # avoid them getting just lost and dropped they are instead written
+                # in fixed format with no compression, which should be just fine
+                # since they have no data anyhow.
+                store.put(
+                    self.pipeline_table_key(table_name, checkpoint_name),
+                    df,
+                )
+            else:
+                store.put(
+                    self.pipeline_table_key(table_name, checkpoint_name),
+                    df,
+                    "table",
+                    complib=complib,
+                )
+            store.flush()
 
     def add_table(self, name, content):
         self._TABLES.add(name)
@@ -605,7 +644,8 @@ class Pipeline:
 
         self.close_open_files()
 
-        self.pipeline_store.close()
+        if not isinstance(self.pipeline_store, Path):
+            self.pipeline_store.close()
 
         self.init_state()
 
@@ -775,12 +815,24 @@ class Pipeline:
         store = self.get_pipeline_store()
 
         if store is not None:
-            df = store[CHECKPOINT_TABLE_NAME]
+            if isinstance(store, Path):
+                df = pd.read_parquet(
+                    store.joinpath(CHECKPOINT_TABLE_NAME, "None.parquet")
+                )
+            else:
+                df = store[CHECKPOINT_TABLE_NAME]
         else:
             pipeline_file_path = config.pipeline_file_path(
                 self.context.get_formatted("pipeline_file_name")
             )
-            df = pd.read_hdf(pipeline_file_path, CHECKPOINT_TABLE_NAME)
+            if pipeline_file_path.endswith(".h5"):
+                df = pd.read_hdf(pipeline_file_path, CHECKPOINT_TABLE_NAME)
+            else:
+                df = pd.read_parquet(
+                    Path(pipeline_file_path).joinpath(
+                        CHECKPOINT_TABLE_NAME, "None.parquet"
+                    )
+                )
 
         # non-table columns first (column order in df is random because created from a dict)
         table_names = [
